Improve ImplementationGuidePackageInterceptor
