+++ matchbox-server/src/main/java/ch/ahdis/matchbox/config/MatchboxJpaConfig.java
@@ -6,10 +6,7 @@ import javax.annotation.Nullable;
 
 import ch.ahdis.matchbox.CliContext;
 import ch.ahdis.matchbox.interceptors.*;
-import ch.ahdis.matchbox.packages.ImplementationGuideProviderR4;
-import ch.ahdis.matchbox.packages.ImplementationGuideProviderR4B;
-import ch.ahdis.matchbox.packages.ImplementationGuideProviderR5;
-import ch.ahdis.matchbox.packages.InstallNpmPackageProvider;
+import ch.ahdis.matchbox.packages.*;
 import ch.ahdis.matchbox.providers.*;
 import ch.ahdis.matchbox.questionnaire.*;
 import ch.ahdis.matchbox.util.MatchboxEngineSupport;
